# Improve performance of large option lists