I've noticed large option lists in selects cause the browser to timeout when rendering selectize. 
Detaching the input, then detaching the children from it, 
then re-attaching the input dramatically improves performance.