@joallard 2016-07-20
Thanks for the patch.

So I've tried this out with a new example in the performance demos and a profiler open. I see no difference with a select with 12k options. Firefox is `5.5s -> 6s` and Chrome `2.5s -> 2.5s`

Would you try branch `pr-1105` at `examples/performance.html`, 2nd example, and figure out if there's a way to test this? Thanks!